data_scrapers/cadweb/requirements.txt
#adafruit-io >= 2.5.0
requests >= 2.25.1
paho-mqtt >= 1.5.1
